requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
requests==2.31.0